+++ src/main/java/blackjack/domain/state/State.java
@@ -3,6 +3,7 @@ package blackjack.domain.state;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import blackjack.domain.game.GameOutcome;
+import blackjack.domain.state.finished.BlackJack;
 
 public interface State {
