+++ src/test/java/blackjack/domain/state/finished/BustTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.state;
+package blackjack.domain.state.finished;
 
 import static blackjack.domain.card.CardNumber.A;
 import static blackjack.domain.card.CardNumber.FOUR;
@@ -7,12 +7,13 @@ import static blackjack.domain.card.CardNumber.KING;
 import static blackjack.domain.card.CardNumber.QUEEN;
 import static blackjack.domain.card.CardPattern.SPADE;
 import static blackjack.domain.game.GameOutcome.DRAW;
-import static blackjack.domain.game.GameOutcome.WIN;
+import static blackjack.domain.game.GameOutcome.LOSE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import blackjack.domain.game.GameOutcome;
+import blackjack.domain.state.State;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.stream.Stream;
@@ -21,13 +22,12 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-class BlackJackTest {
-
+class BustTest {
     @ParameterizedTest
     @DisplayName("자신과 다른 State 비교해 승부 결과를 반환한다.")
     @MethodSource("provideStateAndExpected")
     void compare(State another, GameOutcome expected) {
-        State state = new BlackJack(
+        State state = new Bust(
                 new Cards(Arrays.asList(Card.of(SPADE, KING), Card.of(SPADE, QUEEN), Card.of(SPADE, JACK))));
         assertThat(state.compare(another)).isEqualTo(expected);
     }
@@ -39,9 +39,9 @@ class BlackJackTest {
                 new Cards(Arrays.asList(Card.of(SPADE, KING), Card.of(SPADE, QUEEN), Card.of(SPADE, JACK)));
 
         return Stream.of(
-                Arguments.of(new BlackJack(blackJackCards), DRAW),
-                Arguments.of(new Stay(stayCards), WIN),
-                Arguments.of(new Bust(bustCards), WIN)
+                Arguments.of(new BlackJack(blackJackCards), LOSE),
+                Arguments.of(new Stay(stayCards), LOSE),
+                Arguments.of(new Bust(bustCards), DRAW)
         );
     }
 }
